Restructure serializers into separate files
